requirements.txt
snowflake-snowpark-python==1.11.1
pandas==2.0.3
